Refactor pyon.service.service to pyon.ion.service and pyon.event.event to pyon.ion.event
