+++ pyon/agent/simple_agent.py
@@ -11,7 +11,7 @@ import traceback
 from pyon.core import bootstrap
 from pyon.core.bootstrap import IonObject
 from pyon.core import exception as iex
-from pyon.event.event import EventPublisher
+from pyon.ion.event import EventPublisher
 from pyon.ion.resource import RT, PRED, LCS
 from pyon.util.log import log
 from pyon.util.containers import get_ion_ts
